// game/enums/armor.ts
export enum ArmorType {
    CLOTH = "CLOTH",
    LEATHER = "LEATHER",
    MAIL = "MAIL",
    PLATE = "PLATE"
}

export enum ArmorSlot {
    HELM = "HELM",
    CHEST = "CHEST",
    GLOVES = "GLOVES",
    PANTS = "PANTS",
}


export enum ClothChestName {
    DUSTYROBES = "Dusty Robes",
    DIRTYTUNIC = "Dirty Tunic",
    CLEANSHIRT = "Clean Shirt",
    BUGCLOTHCHEST = "Bug Infested Cloth Chest"
}

export enum LeatherChestName {
    BUGINFESTEDLEAHTERCHEST = "Bug Infested Leather Chest", 
    TATTEREDCHEST = 'Tattered Chest',
    JERKIN = "Jerkin",
    GOBLINSOLDIERLEATHERJACKET = "Goblin Soldier's Leather Jacket",
}

export enum LeatherGlovesName {
    BUGINFESTEDLEATHERGLOVES = "Bug Infested Leather Gloves",
    SHREDDEDGLOVES = "Shredded Gloves",
    DUSTYGLOVES = "Dusty Gloves",
    DWARFMININGGLOVES = "Dwarf Mining Gloves",
    THIEFGLOVES = "Theif Gloves",
}

export enum LeatherHelmName {
    BUGINFESTEDLEATHERHELM = "Bug Infested Leather Helm",
    LEATHERHEADBAND = "Leather Headband",
    TRIBALWARHELM = "Tribal War Helm",
}

export enum LeatherPantsName {
    BUGINFESTEDLEATHERPANTS = "Bug Infested Leather Pants",
    DANCERSCHAPS = "Dancers Chaps",
    DIRTYPANTS = "Dirty Pants",
    PATCHEDPANTS = "Patched Pants",
    MUSHROOMINFECTEDPANTS = "Mushroom Infected Pants"
}

export enum MailChestName {
    BUGINFESTEDCHAINCHEST = "Bug Infested Chain Chest",
    RUSTYCHAINCHEST = "Rusty Chain Chest", 
    SOLDIERCHAINCHEST = "Soldier Chain Chest",
    ELVINSCHAINCHEST = "Elvins Chain Chest",
    OGRECAPTAINCHEST = "Ogre Captain Chest"
}

export enum MailGlovesName {
    BUGINFESTEDCHAINGLOVES = "Bug Infested Chain Gloves",
    CHAINGLOVES = "Chain Gloves",
    ELVENCHAINGLOVES = "Elven Chain Gloves",
    SQUIRETRAININGGLOVES = "Squire Training Gloves",
    SPIDERSILKGLOVES = "Spider Silk Gloves"
}

export enum MailHelmName {
    BUGINFESTEDMAILHELM = "Bug Infested Mail Helm",
    BROKENCOPPERHELM = "Broken Copper Helm",
    SKULLCAP = "Skullcap",
    GLISTENINGCOIF = "Glistening Coif",
}

export enum MailPantsName {
    BUGINFESTEDMAILPANTS = "Bug Infested Mail Pants",
    COPPERPANTS = "Copper Pants",
    RUSTYPAINTS = "Rusty Pants",
    GOBLINSOLDIERPANTS = "Goblin Soldier Pants",
    BLACKSMITHSONSPANTS = "Blacksmith's Son Pants"
}